wip replace platformScaffold with platformTabScaffold when bottomNavbar is needed
